graphics/requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
pillow==11.1.0 
